+++ src/demos/Demos.ImGuiBackend.GlfwHexa/Services/App.cs
@@ -2,6 +2,7 @@
 using Demos.ImGuiBackend.GlfwHexa.Utils;
 using Detach.GlfwExtensions;
 using Detach.ImGuiBackend.GlfwHexa;
+using Detach.Metrics;
 using Hexa.NET.ImGui;
 using Silk.NET.GLFW;
 using Silk.NET.OpenGL;
@@ -23,7 +24,8 @@ public sealed class App
 	private readonly unsafe WindowHandle* _window;
 	private readonly GlfwInput _glfwInput;
 	private readonly ImGuiController _imGuiController;
-	private readonly PerformanceMeasurement _performanceMeasurement;
+	private readonly FrameCounter _frameCounter;
+	private readonly HeapAllocationCounter _heapAllocationCounter;
 	private readonly InputDebugWindow _inputDebugWindow;
 	private readonly KeyboardInputWindow _keyboardInputWindow;
 	private readonly MouseInputWindow _mouseInputWindow;
@@ -40,7 +42,8 @@ public sealed class App
 		WindowHandle* window,
 		GlfwInput glfwInput,
 		ImGuiController imGuiController,
-		PerformanceMeasurement performanceMeasurement,
+		FrameCounter frameCounter,
+		HeapAllocationCounter heapAllocationCounter,
 		InputDebugWindow inputDebugWindow,
 		KeyboardInputWindow keyboardInputWindow,
 		MouseInputWindow mouseInputWindow,
@@ -52,7 +55,8 @@ public sealed class App
 		_window = window;
 		_glfwInput = glfwInput;
 		_imGuiController = imGuiController;
-		_performanceMeasurement = performanceMeasurement;
+		_frameCounter = frameCounter;
+		_heapAllocationCounter = heapAllocationCounter;
 		_inputDebugWindow = inputDebugWindow;
 		_keyboardInputWindow = keyboardInputWindow;
 		_mouseInputWindow = mouseInputWindow;
@@ -93,7 +97,7 @@ public sealed class App
 		if (_frameTime > _maxMainDelta)
 			_frameTime = _maxMainDelta;
 
-		_performanceMeasurement.UpdateFrameTime(mainStartTime, _frameTime);
+		_frameCounter.Update(mainStartTime, _frameTime);
 
 		_currentTime = mainStartTime;
 		_accumulator += _frameTime;
@@ -118,7 +122,7 @@ public sealed class App
 
 		ImGui.ShowDemoWindow();
 
-		_performanceMeasurement.UpdateAllocatedBytes(GC.GetAllocatedBytesForCurrentThread());
+		_heapAllocationCounter.UpdateAllocatedBytesForCurrentThread();
 
 		_inputDebugWindow.Render();
 		_keyboardInputWindow.Render();