+++ src/demos/Demos.ImGuiBackend.GlfwHexa/Services/Ui/PerformanceWindow.cs
@@ -1,15 +1,18 @@
 ﻿using Detach;
+using Detach.Metrics;
 using Hexa.NET.ImGui;
 
 namespace Demos.ImGuiBackend.GlfwHexa.Services.Ui;
 
 public sealed class PerformanceWindow
 {
-	private readonly PerformanceMeasurement _performanceMeasurement;
+	private readonly FrameCounter _frameCounter;
+	private readonly HeapAllocationCounter _heapAllocationCounter;
 
-	public PerformanceWindow(PerformanceMeasurement performanceMeasurement)
+	public PerformanceWindow(FrameCounter frameCounter, HeapAllocationCounter heapAllocationCounter)
 	{
-		_performanceMeasurement = performanceMeasurement;
+		_frameCounter = frameCounter;
+		_heapAllocationCounter = heapAllocationCounter;
 	}
 
 	public void Render()
@@ -18,13 +21,13 @@ public sealed class PerformanceWindow
 		{
 			ImGui.SeparatorText("Rendering");
 
-			ImGui.Text(Inline.Utf8($"{_performanceMeasurement.Fps} FPS"));
-			ImGui.Text(Inline.Utf8($"Frame time: {_performanceMeasurement.FrameTime:0.0000} s"));
+			ImGui.Text(Inline.Utf8($"{_frameCounter.FrameCountPreviousSecond} FPS"));
+			ImGui.Text(Inline.Utf8($"Frame time: {_frameCounter.CurrentFrameTime:0.0000} s"));
 
 			ImGui.SeparatorText("Allocations");
 
-			ImGui.Text(Inline.Utf8($"Allocated: {_performanceMeasurement.AllocatedBytes:N0} bytes"));
-			ImGui.Text(Inline.Utf8($"Since last update: {_performanceMeasurement.AllocatedBytesSinceLastUpdate:N0} bytes"));
+			ImGui.Text(Inline.Utf8($"Allocated: {_heapAllocationCounter.AllocatedBytes:N0} bytes"));
+			ImGui.Text(Inline.Utf8($"Since last update: {_heapAllocationCounter.AllocatedBytesSinceLastUpdate:N0} bytes"));
 
 			for (int i = 0; i < GC.MaxGeneration + 1; i++)
 				ImGui.Text(Inline.Utf8($"Gen{i}: {GC.CollectionCount(i)} times"));