+++ src/demos/Demos.ImGuiBackend.GlfwImGuiNET/Services/Ui/PerformanceWindow.cs
@@ -1,17 +1,18 @@
 ﻿using Detach;
+using Detach.Metrics;
 using ImGuiNET;
 
 namespace Demos.ImGuiBackend.GlfwImGuiNET.Services.Ui;
 
 public sealed class PerformanceWindow
 {
-	private long _previousAllocatedBytes;
+	private readonly FrameCounter _frameCounter;
+	private readonly HeapAllocationCounter _heapAllocationCounter;
 
-	private readonly PerformanceMeasurement _performanceMeasurement;
-
-	public PerformanceWindow(PerformanceMeasurement performanceMeasurement)
+	public PerformanceWindow(FrameCounter frameCounter, HeapAllocationCounter heapAllocationCounter)
 	{
-		_performanceMeasurement = performanceMeasurement;
+		_frameCounter = frameCounter;
+		_heapAllocationCounter = heapAllocationCounter;
 	}
 
 	public void Render()
@@ -20,15 +21,13 @@ public sealed class PerformanceWindow
 		{
 			ImGui.SeparatorText("Rendering");
 
-			ImGui.Text(Inline.Utf16($"{_performanceMeasurement.Fps} FPS"));
-			ImGui.Text(Inline.Utf16($"Frame time: {_performanceMeasurement.FrameTime:0.0000} s"));
+			ImGui.Text(Inline.Utf16($"{_frameCounter.FrameCountPreviousSecond} FPS"));
+			ImGui.Text(Inline.Utf16($"Frame time: {_frameCounter.CurrentFrameTime:0.0000} s"));
 
 			ImGui.SeparatorText("Allocations");
 
-			long allocatedBytes = GC.GetAllocatedBytesForCurrentThread();
-			ImGui.Text(Inline.Utf16($"Allocated: {allocatedBytes:N0} bytes"));
-			ImGui.Text(Inline.Utf16($"Since last update: {allocatedBytes - _previousAllocatedBytes:N0} bytes"));
-			_previousAllocatedBytes = allocatedBytes;
+			ImGui.Text(Inline.Utf16($"Allocated: {_heapAllocationCounter.AllocatedBytes:N0} bytes"));
+			ImGui.Text(Inline.Utf16($"Since last update: {_heapAllocationCounter.AllocatedBytesSinceLastUpdate:N0} bytes"));
 
 			for (int i = 0; i < GC.MaxGeneration + 1; i++)
 				ImGui.Text(Inline.Utf16($"Gen{i}: {GC.CollectionCount(i)} times"));